fix: 5-sqrt_recursion.c task

+++ 0x08-recursion/main.h
@@ -9,7 +9,7 @@ int _inner_strlen(char *s, int n);
 int factorial(int n);
 int _pow_recursion(int x, int y);
 int _sqrt_recursion(int n);
-int _inner_sqrt(int n, int m);
+int _inner_sqrt(unsigned long n, unsigned long m);
 int is_prime_number(int n);
 int _is_prime_number(int n, int m);
